new algorithm to find irreducible sector
